+++ apple2/src/fn_fuji/fuji_appkey_open.c
@@ -2,7 +2,8 @@
 #include "fujinet-fuji.h"
 #include "fujinet-bus-apple2.h"
 
-uint8_t fuji_appkey_open(AppKeyOpen *buffer)
+// Open app-key, returns error status
+bool fuji_appkey_open(AppKeyOpen *buffer)
 {
 	sp_error = sp_get_fuji_id();
 	if (sp_error <= 0) {