+++ apple2/src/fn_fuji/fuji_base64_decode_output.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include "fujinet-fuji.h"
 
-uint8_t fuji_base64_decode_output(char *s, uint16_t len)
+bool fuji_base64_decode_output(char *s, uint16_t len)
 {
-    return 1;
+    return true;
 }