+++ apple2/src/fn_fuji/fuji_base64_encode_compute.c
@@ -3,10 +3,10 @@
 #include "fujinet-fuji.h"
 #include "fujinet-bus-apple2.h"
 
-uint8_t fuji_base64_encode_compute()
+bool fuji_base64_encode_compute()
 {
     // send CTRL command: 0xCF
     // PAYLOAD: N/A
 
-    return sp_control(0, 0xCF);
+    return sp_control(0, 0xCF) == 0;
 }