+++ apple2/src/fn_fuji/fuji_hash_compute.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include "fujinet-fuji.h"
 
-uint8_t fuji_hash_compute(uint8_t type)
+bool fuji_hash_compute(uint8_t type)
 {
-    return 0;
+    return true;
 }