+++ commodore/src/fn_fuji/fuji_appkey_read.c
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include "fujinet-fuji.h"
 
-uint8_t fuji_appkey_read(AppKeyRead *buffer)
+bool fuji_appkey_read(AppKeyRead *buffer)
 {
 	return 0;
 }