+++ commodore/src/fn_fuji/fuji_base64_decode_compute.c
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include "fujinet-fuji.h"
 
-uint8_t fuji_base64_decode_compute()
+bool fuji_base64_decode_compute()
 {
 	return 0;
 }