+++ commodore/src/fn_fuji/fuji_base64_decode_input.c
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include "fujinet-fuji.h"
 
-uint8_t fuji_base64_decode_input(char *s, uint16_t len)
+bool fuji_base64_decode_input(char *s, uint16_t len)
 {
 	return 0;
 }