+++ commodore/src/fn_fuji/fuji_base64_encode_length.c
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include "fujinet-fuji.h"
 
-uint8_t fuji_base64_encode_length(unsigned long *len)
+bool fuji_base64_encode_length(unsigned long *len)
 {
 	return 0;
 }