+++ commodore/src/fn_fuji/fuji_hash_input.c
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include "fujinet-fuji.h"
 
-uint8_t fuji_hash_input(char *s, uint16_t len)
+bool fuji_hash_input(char *s, uint16_t len)
 {
 	return 0;
 }