+++ commodore/src/fn_fuji/fuji_hash_length.c
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include "fujinet-fuji.h"
 
-uint8_t fuji_hash_length(uint8_t mode)
+bool fuji_hash_length(uint8_t mode)
 {
 	return 0;
 }